Attempted to fix broken reverse proxy config after
upgrade of LXC server from bionic to jammy.
See https://github.com/kommserv/netbox/issues/2

Also some housekeeping and replacement of hard-coded
stuff with variables.

Note that vhost config on the reverse proxy is
not handled by this role.
In this role I experimented with adding a vhost
config on the container itself, thus effectively
bouncing every request through two proxies
(this works for some other services, for example
TinyTinyRSS and Wallabag), but did not work for
Netbox.

Code in this commit is very much in flux,
no guarantees anything is working properly.